+++ test/Telegram.Bot.Tests.Integ/Admin Bot/SupergroupAdminBotTests.cs
@@ -11,38 +11,61 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
 {
     [Collection(Constants.TestCollections.SupergroupAdminBots)]
     [TestCaseOrderer(Constants.TestCaseOrderer, Constants.AssemblyName)]
-    public class SupergroupAdminBotTests : IClassFixture<AdminBotTestFixture>
+    public class SupergroupAdminBotTests : IClassFixture<SupergroupAdminBotTestsFixture>
     {
-        private readonly AdminBotTestFixture _classFixture;
+        private readonly SupergroupAdminBotTestsFixture _classFixture;
 
-        private readonly TestsFixture _fixture;
+        private ITelegramBotClient BotClient => _classFixture.TestsFixture.BotClient;
 
-        private ITelegramBotClient BotClient => _fixture.BotClient;
-
-        public SupergroupAdminBotTests(TestsFixture testsFixture, AdminBotTestFixture classFixture)
+        public SupergroupAdminBotTests(SupergroupAdminBotTestsFixture classFixture)
         {
-            _fixture = testsFixture;
             _classFixture = classFixture;
-            _classFixture.Chat = _fixture.SupergroupChat;
         }
 
         #region 1. Changing Chat Title
 
-        [OrderedFact(DisplayName = FactTitles.ShouldSetChatTitle)]
+        [OrderedFact("Should set chat title")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatTitle)]
         public async Task Should_Set_Chat_Title()
         {
             await BotClient.SetChatTitleAsync(
                 chatId: _classFixture.Chat.Id,
-                title: _classFixture.ChatTitle
+                title: "Test Chat Title"
             );
         }
 
         #endregion
 
-        #region 2. Changing Chat Description
+        #region 2. Changing Chat default permissions
 
-        [OrderedFact(DisplayName = FactTitles.ShouldSetChatDescription)]
+        [OrderedFact("Should set new default permissions")]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatPermissions)]
+        public async Task Should_Set_New_Default_Permissions()
+        {
+            ChatPermissions newDefaultPermissions = new ChatPermissions()
+            {
+                CanInviteUsers = false,
+                CanSendMediaMessages = true,
+                CanChangeInfo = true,
+                CanSendMessages = true,
+                CanPinMessages = true,
+                CanSendPolls = false,
+                CanSendOtherMessages = false,
+                CanAddWebPagePreviews = false
+            };
+
+            await BotClient.SetChatPermissionsAsync(_classFixture.Chat.Id, newDefaultPermissions);
+            Chat supergroup = await BotClient.GetChatAsync(_classFixture.Chat.Id);
+            ChatPermissions setChatPermissions = supergroup.Permissions;
+
+            Asserts.JsonEquals(newDefaultPermissions, setChatPermissions);
+        }
+
+        #endregion
+
+        #region 3. Changing Chat Description
+
+        [OrderedFact("Should set chat description")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatDescription)]
         public async Task Should_Set_Chat_Description()
         {
@@ -52,7 +75,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             );
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldDeleteChatDescription)]
+        [OrderedFact("Should delete chat description")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatDescription)]
         public async Task Should_Delete_Chat_Description()
         {
@@ -65,13 +88,13 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
 
         #endregion
 
-        #region 3. Pinning Chat Description
+        #region 4. Pinning Chat Description
 
-        [OrderedFact(DisplayName = FactTitles.ShouldPinMessage)]
+        [OrderedFact("Should pin chat message")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.PinChatMessage)]
         public async Task Should_Pin_Message()
         {
-            Message msg = await _fixture.SendTestInstructionsAsync("🧷 This message will be pinned shortly!");
+            Message msg = await _classFixture.TestsFixture.SendTestInstructionsAsync("🧷 This message will be pinned shortly!");
 
             await BotClient.PinChatMessageAsync(
                 chatId: _classFixture.Chat.Id,
@@ -82,7 +105,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             _classFixture.PinnedMessage = msg;
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldGetChatPinnedMessage)]
+        [OrderedFact("Should get chat's pinned message")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.GetChat)]
         public async Task Should_Get_Chat_Pinned_Message()
         {
@@ -95,14 +118,14 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             ));
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldUnpinMessage)]
+        [OrderedFact("Should unpin chat message")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.UnpinChatMessage)]
         public async Task Should_Unpin_Message()
         {
             await BotClient.UnpinChatMessageAsync(_classFixture.Chat.Id);
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldGetChatWithNoPinnedMessage)]
+        [OrderedFact("Should get the chat info without a pinned message")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.GetChat)]
         public async Task Should_Get_Chat_With_No_Pinned_Message()
         {
@@ -113,9 +136,9 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
 
         #endregion
 
-        #region 4. Changing Chat Photo
+        #region 5. Changing Chat Photo
 
-        [OrderedFact(DisplayName = FactTitles.ShouldSetChatPhoto)]
+        [OrderedFact("Should set chat photo")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatPhoto)]
         public async Task Should_Set_Chat_Photo()
         {
@@ -128,14 +151,14 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             }
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldDeleteChatPhoto)]
+        [OrderedFact("Should delete chat photo")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.DeleteChatPhoto)]
         public async Task Should_Delete_Chat_Photo()
         {
             await BotClient.DeleteChatPhotoAsync(_classFixture.Chat.Id);
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldThrowOnDeletingChatDeletedPhoto)]
+        [OrderedFact("Should throw exception in deleting chat photo with no photo currently set")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.DeleteChatPhoto)]
         public async Task Should_Throw_On_Deleting_Chat_Deleted_Photo()
         {
@@ -147,48 +170,18 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             Assert.Equal("Bad Request: CHAT_NOT_MODIFIED", e.Message);
         }
 
-        /// <summary>
-        /// If chat had a photo before, reset the photo back.
-        /// </summary>
-        [OrderedFact(DisplayName = FactTitles.ShouldResetOldChatPhoto)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatPhoto)]
-        public async Task Should_Reset_Old_Chat_Photo_If_Existed()
-        {
-            // "Chat.Photo" might be null if there is no photo currently set
-            string previousChatPhotoId = _classFixture.Chat.Photo?.BigFileId;
-            if (previousChatPhotoId == default)
-            {
-                // chat didn't have a photo
-                return;
-            }
-
-            using (Stream photoStream = new MemoryStream())
-            {
-                // pass photo's file_id, prepare file for download, and download the file into memroy
-                await BotClient.GetInfoAndDownloadFileAsync(previousChatPhotoId, photoStream);
-
-                // need to set position of memory stream back to its start so next method reads photo stream from the beginning
-                photoStream.Position = 0;
-
-                await BotClient.SetChatPhotoAsync(
-                    chatId: _classFixture.Chat.Id,
-                    photo: photoStream
-                );
-            }
-        }
-
         #endregion
 
-        #region 5. Chat Sticker Set
+        #region 6. Chat Sticker Set
 
-        [OrderedFact(DisplayName = FactTitles.ShouldThrowOnSettingChatStickerSet)]
+        [OrderedFact("Should throw exception when trying to set sticker set for a chat with less than 100 members")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatStickerSet)]
         public async Task Should_Throw_On_Setting_Chat_Sticker_Set()
         {
             const string setName = "EvilMinds";
 
             ApiRequestException exception = await Assert.ThrowsAnyAsync<ApiRequestException>(() =>
-                _fixture.BotClient.SetChatStickerSetAsync(_classFixture.Chat.Id, setName)
+                BotClient.SetChatStickerSetAsync(_classFixture.Chat.Id, setName)
             );
 
             // ToDo: Create exception type
@@ -197,34 +190,5 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         }
 
         #endregion
-
-        private static class FactTitles
-        {
-            public const string ShouldSetChatTitle = "Should set chat title";
-
-            public const string ShouldSetChatDescription = "Should set chat description";
-
-            public const string ShouldDeleteChatDescription = "Should delete chat description";
-
-            public const string ShouldPinMessage = "Should pin chat message";
-
-            public const string ShouldGetChatPinnedMessage = "Should get chat's pinned message";
-
-            public const string ShouldUnpinMessage = "Should unpin chat message";
-
-            public const string ShouldGetChatWithNoPinnedMessage = "Should get the chat info without a pinned message";
-
-            public const string ShouldSetChatPhoto = "Should set chat photo";
-
-            public const string ShouldDeleteChatPhoto = "Should delete chat photo";
-
-            public const string ShouldThrowOnDeletingChatDeletedPhoto =
-                "Should throw exception in deleting chat photo with no photo currently set";
-
-            public const string ShouldResetOldChatPhoto = "Should reset the same old chat photo if existed";
-
-            public const string ShouldThrowOnSettingChatStickerSet =
-                "Should throw exception when trying to set sticker set for a chat with less than 100 members";
-        }
     }
 }